Moved DatabaseManager from BarcodeAPI folder to Util folder since the database is not used ONLY for the barcode.
Modified DatabaseManager so that the constructor accepts connection string. It was hardcoded into the constructor but that is not versatile if the developer wants to change the settings. Added constants in the constants module for the database information.

Modified BarcodeManager so that it accepts a DatabaseManager instance instead of initializing one. This is because the database is used in other places besides the BarcodeManager and having multiple instances of DatabaseManager will be inefficient. (Each DatabaseManager instance opens a SQL connection which utilizes more resources)

Made changes to the database manager because in the psycopg2 usage documentation, it says to NEVER EVER NEVER EVER (with that much emphasis) pass arguments using the % or + operator. Instead, the execute statement accepts arguments as parameters. I'm assuming this is because it will esacpe it for you, not sure though.
