+++ BarcodeAPI/barcodeManager.py
@@ -1,11 +1,11 @@
 import urllib.request  # For internet operations
 import json  # To internet JSON data format
-from .databaseManager import DatabaseManager
+from Util.databaseManager import DatabaseManager
 
 
 class BarcodeManager:
-    def __init__(self):
-        self.dbManager = DatabaseManager()
+    def __init__(self, dbManager):
+        self.dbManager = dbManager
 
     def AddItemToInventory(self, barcode):
         cachedItem = self.dbManager.GetCachedUPCItem()
