+++ Tests/databaseManager.Tests.py
@@ -1,9 +1,12 @@
-from BarcodeAPI.databaseManager import DatabaseManager
+from Util.databaseManager import DatabaseManager
+from Util import constants
 
 # Test init for connecting to DB
-dbManager = DatabaseManager()
+dbManager = DatabaseManager(constants.dbDatabase, constants.dbUsername, constants.dbPassword, constants.dbHost, constants.dbPort)
 # Test Cache UPC retrieval method
-cachedUPC = dbManager.GetCachedItem("070360002372")
+cachedUPC = dbManager.GetCachedUPCItem("070360002372")
+
+dbManager.GetItemFromInventory(2)
 
 
 print("")
